fea/scheduler

+++ database/migrations/2020_10_13_045757_add_post_to_schedulers_table.php
@@ -27,8 +27,10 @@ class AddPostToSchedulersTable extends Migration
      */
     public function down()
     {
-        Schema::table('schedulers', function (Blueprint $table) {
-            $table->dropColumn('post_id');
-        });
+        if (Schema::hasColumn('schedulers', 'post_id')) {
+            Schema::table('schedulers', function (Blueprint $table) {
+                $table->dropColumn('post_id');
+            });
+        }
     }
 }
